[#28] Hold HashMap dictionary in memory
