+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
-* Optimize `cli` to hold mapped dictinary in memory.
+* [#28](https://github.com/willbasky/TibetCli/issues/28):
+  Hold `HashMap` dictinaries in memory. Preload it before a first query.
+* Move unused functions to Sandbox module.
+
 
 ## [0.3] - 2018-10-01
