+++ tibet.cabal
@@ -43,7 +43,7 @@ library
   build-depends:       base
                      , ansi-terminal
                      , bytestring
-                     , containers
+                     , deepseq
                      , directory
                      , gitrev
                      , neat-interpolation ^>= 0.3.2.2
@@ -51,6 +51,7 @@ library
                      , path
                      , path-io >= 1.4.0
                      , text
+                     , unordered-containers
 
   default-language:    Haskell2010
   default-extensions:  DeriveGeneric
